+++ app/src/main/AndroidManifest.xml
@@ -19,6 +19,12 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <service
+            android:name=".SendDataToServer"
+            android:exported="false">
+
+        </service>
+
 
     </application>
